Fix MCP server installation tracking
